+++ packages/koa-change-case/lib/converter/index.ts
@@ -1,4 +1,4 @@
 export { default as MiddlewareFactory } from "./middleware-factory";
-export { default as defaultMiddlewareFactoryProvider } from "./default-converter-manager-factory";
+export { default as createDefaultConverterManager } from "./create-default-converter-manager";
 export { default as toMiddlewareFactory } from "./to-middleware-factory";
-export * from "./converters";
+export * from "./middleware-factories";
