+++ packages/koa-change-case/lib/index.ts
@@ -2,3 +2,4 @@ export * from "convable";
 export * from "./exchanger";
 export * from "./convert-processor";
 export * from "./converter";
+export * from "./change-case";
